requirements.txt
alibi-detect
mlserver
pytest
pytest-asyncio
black
pylint
mypy
types-requests
requests-mock